Earth Spell Mechanins
Working on getting the scripts to work, but the idea is that it spawns an invisible cube in the same size as the earth object cube first. It then checks for collision. If there is collision, it will not let you spawn the earth object cube, but if there isnt, it will delete the invisible placeholder cube and then drop in the real one. Close but not quite there yet.