First Milestone Draft (#17)
* Basic Blocking and Character Integration

Blocked out the basic layout of the level and integrated the basic player and main camera prefab.

* Created a Prefab and Script for Moving Platform

There is a terst level so you can see how it moves, there are also colliders to test out with the Human tester we have.

* Compile Error Fix

I made an oopsie!

* Updates to Player Prefab

* Added TutorialEventSystem.cs

* Just Push

* Update

* Anotha One

* Started work on Circuit Puzzle

* Established CircuitPuzzleController.cs

* Setup TurretEnemy (Basic Frame)

* Circuit Puzzle Switches Working

Got the switches working to rotate the puzzle.

* Updated Tutorial Scene

* Rings Rotate Smoothly

Got slerping done right.

* Updated CircuitPuzzleController.cs

Updated the Date Last Modified in CircuitPuzzleController.cs.

* More changes

* Modified CircuitPuzzleController.cs

Worked on it more.

* Code Comments

Commented on my code because everyone be hatin! XD

* Edits to Circuit Puzzle Prefab

* Extended Side Rooms

* Added Player and Platform Functionality

Added class to allow for the player to not have to move when on a moving platform.

* MovingPlatform Development Stage Changes

Added 2 platforms (Vertical and Diagonal) for testing purposes.

* Updated right room

* Update left room platforms

* Pushing Updates for Jumping

* Blocked Out Key Rooms + Worked more on CircuitPuzzleController.cs

Modified Brenda's changes for the right key room, blocked out the left key room.

* Updated Key Room Layout

Made the two key rooms easier in terms of their difficulty.

* Update

* Modified Jumping

Rid Human of bounciness.

* Integrated "Badman" Enemy

* Added walljump

* Added Gyms

Added gyms for Human Movement and Ferrox Movement. Ferrox meovement gym may need a bit more work.