Pretty Close
So array gets populated with positions. (WindSpellUse.cs)
Positions get transffered to WindSpellMover.cs
...But the positions are totally wrong.
I think the problem is in the code for WindSpellUse.cs that chooses the positions to spawn the objects. Only confusing thing is that in Unity the trgtObjs look like the have realistic coordinates. Figure it out Evan.