Very Close to Working
There is a problem with the way the entire player is implemented right now that should be addressed, and needs to be changed in order for the checkpoint system (or any teleport system) to work properly. The parent game object for the player doesn't move, only the children, which is sloppy because the parent just sits at the spawn like a vegetable as the human moves around. This does not seem like an intended behavior and once changed, I can make the respawn system work.