SortedJumpScripts
The jump script that we want to be using is attached to the player, and the redundant ones have been placed in an "unused scripts" folder. The Checkpoint and Health scene has yet to be repaired, as it needs 3D surfaces for the player to walk on (I will do this soonish). I also updated the player prefab and the camera prefab, as they are both important in relation to controlling the player.