Added Surfaces to Scene 0
I'm trying to fix the jump/ ground check not working when moving from a slope to a flat surface. I have discovered that this is due to an unintended behavior from the "OnTriggerExit" function in the ground check script. When the detection box for ground check leaves one platform, despite instantly detecting another, it will perform its action, setting the "isColliding" boolean to false, Now that I know why this happens, I can work towards fixing it.