Side Collisions Work Consistently
Side collisions work aside from being violently impaled. 

A new public value maxClimbableSlope allows the max angle that can be considered the "ground" and not a wall. It is currently set to 70 deg.

A bug with the gravity if statement has been fixed. [Line 134] 

Issues: Jumping into a ceiling gets you head stuck in the clouds. An attempt to fix it has been added to [Line 134] "|| ceilingPos != Vector3.zero" but it still gets stuck

The left side of the ground checker does not seem to work on the slope in the middle.