Fixed a Bug
Found a bug where if you were sliding down a wall jump wall that ended, you would continue to slide through the air as if the wall was still there. I added a ray that is casted while the player is sliding to make sure they are still on the wall. I could not use the pre-existing rays, as they are not casted unless the player is moving into the wall, which is not needed to wall slide.