FinalPushBeforeAudio
Left out all the mana stuff because it just would not behave. I might have to scrap most of what I have. In any event, this should be pretty solid. It SHOULD be on the right version of Unity but left me know if there is any issues and I can fix it.