Update: Read Description
So I did more than place the killbox in the scene. At the moment, I don't know how the whole death and health system functions so it isn't synced up but the killbox does have a collider that is adjusted to the length and shape of the demo level.

The more important changes. 

I adjusted the speed of both Yumi and the Yokai (see what I did there?). Yumi now has a slightly faster horizontal speed, giving the player a bit more control and freedom when maneuvering platforms and in the air. The yokai's horizontal speed has also been greatly accelerated, which creates a more nimble and quickfooted feel. Mechanically, the increase in acceleration offers the player more freedoms in jumping. 

I have adjusted the locations of some of the leaf platforms. You will see that they no longer connect to any branches. These positions feel more natural.