Updated Flying Enemy animations
I attached the new Flying Enemy animations, and I added to the Flying Enemy script. I created a Flip function in the enemy script that will make the animation flip about the y-axis to face the correct direction while it is flying. So, now the bat animation will face left when the enemy is moving left and vice versa.