Fixing Windspell Bug
There was a mismatch between the name of a script in the editor, and the name of the script as referenced by the script itself. For whatever reason, git isn't catching renames of files in editor, so instead, I changed the actual filename to match the name in editor. This should fix the issue.